refac: TurnService 만들어서 차례 table 관리

+++ src/main/java/chess/service/BoardService.java
@@ -5,52 +5,75 @@ import chess.domain.board.BoardDAO;
 import chess.domain.board.BoardFactory;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
+import chess.exception.TakeTurnException;
 
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
 
 public class BoardService {
-    public static Map<String, Object> parseEmptyBoard() {
+    private BoardDAO boardDAO;
+
+    public BoardService() {
+        boardDAO = new BoardDAO();
+    }
+
+    public Map<String, Object> receiveEmptyBoard() {
         Board board = BoardFactory.createBoard();
         return createBoardModel(board);
     }
 
-    public static Map<String, Object> parseInitializedBoard(final Board board, final BoardDAO boardDAO) throws SQLException {
-        boardDAO.deletePieces();
-        for (Position position : board.getBoard().keySet()) {
-            boardDAO.placePiece(board, position);
-        }
-        Map<String, Object> model = createBoardModel(board);
-        model.put("turn", board.getCurrentTurn() + "가 먼저 시작합니다.");
-        return model;
-    }
-
-    public static Board receiveLoadedBoard(final BoardDAO boardDAO) throws SQLException {
-        return new Board(boardDAO.findAllPieces());
-    }
-
-    public static Map<String, Object> parseLoadedBoard(final BoardDAO boardDAO) throws SQLException {
-        return createBoardModel(receiveLoadedBoard(boardDAO));
-    }
-
-    public static Map<String, Object> receiveMovedBoard(final Board board, final BoardDAO boardDAO) throws SQLException {
-        Map<String, Object> model = createBoardModel(board);
-        model.put("turn", board.getCurrentTurn() + "의 차례입니다.");
+    public Map<String, Object> receiveInitializedBoard() throws SQLException {
+        Board board = BoardFactory.createBoard();
+        board.initialize();
 
         boardDAO.deletePieces();
         for (Position position : board.getBoard().keySet()) {
             boardDAO.placePiece(board, position);
         }
+
+
+        Map<String, Object> model = createBoardModel(board);
+        model.put("turn",  "WHITE 가 먼저 시작합니다.");
         return model;
     }
 
-    public static Map<String, Object> createBoardModel(final Board board) {
+    public Map<String, Object> receiveLoadedBoard(TurnService turnService) throws SQLException {
+        Board board = new Board(boardDAO.findAllPieces());
+        Map<String, Object> model = createBoardModel(board);
+        model.put("turn", turnService.getCurrentTurn() + "차례 입니다.");
+        return model;
+    }
+
+    public Map<String, Object> receiveMovedBoard(final String fromPiece, final String toPiece, final TurnService turnService) throws SQLException {
+        Board board = new Board(boardDAO.findAllPieces());
+        Piece piece = board.findBy(Position.of(fromPiece));
+        if (!piece.isSameTeam(turnService.getCurrentTurn())) {
+            throw new TakeTurnException("체스 게임 순서를 지켜주세요.");
+        }
+        board.move(fromPiece, toPiece);
+        boardDAO.deletePieces();
+        for (Position position : board.getBoard().keySet()) {
+            boardDAO.placePiece(board, position);
+        }
+
+        Map<String, Object> model = createBoardModel(board);
+        turnService.updateTurn();
+        model.put("turn", turnService.getCurrentTurn() + "차례입니다.");
+        return model;
+    }
+
+    private Map<String, Object> createBoardModel(final Board board) {
         Map<String, Object> model = new HashMap<>();
         Map<Position, Piece> rawBoard = board.getBoard();
         for (Position position : rawBoard.keySet()) {
             model.put(position.toString(), rawBoard.get(position).getName());
         }
         return model;
+    }
+
+    public boolean isFinished() throws SQLException {
+        Board board = new Board(boardDAO.findAllPieces());
+        return board.isFinished();
     }
 }
